requirements.txt
easydict
einops
lxml
matplotlib
mayavi
numpy
opencv-python
pillow
tensorboardX
tqdm
xvfbwrapper
wandb
